% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AusWild_geom.R
\docType{data}
\name{AusWild_geom}
\alias{AusWild_geom}
\title{Geometry file for Australian Wildfire data}
\format{
\describe{
\item{AusWild_geom}{A sf data frame containing the geometry of the (7901) polygons that make up the spatial domain.}
}
}
\usage{
data(AusWild_geom)
}
\description{
Data used by Cisneros et al. (2023+) for modelling extreme wildfires in Australia
with eGPD graph convolutional neural networks
}
\details{
See \code{help("AusWild")} for relevant response and covariates.
}
\examples{

data("AusWild_geom")

}
\references{
Cisneros, D., Richards, J., Dahal, A., Lombardo, L., and Huser, R. (2023+), \emph{Deep learning-based graphical regression for jointly moderate and extreme Australian wildfires.}. (\href{}{In draft}).
}
\keyword{datasets}
