% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eGPD.R
\name{eGPD}
\alias{eGPD}
\alias{peGPD}
\alias{qeGPD}
\alias{reGPD}
\title{The extended GPD distribution}
\usage{
peGPD(y, sigma, kappa, xi, log = F)

qeGPD(prob, sigma, kappa, xi)

reGPD(n, sigma, kappa, xi)
}
\arguments{
\item{y}{scalar quantile.}

\item{sigma}{scalar scale parameter.}

\item{kappa}{scalar location parameter.}

\item{xi}{scalar shape parameter.}

\item{log}{logical; if \code{TRUE}, probabilities are given as \code{log(prob)}.}

\item{prob}{scalar probability.}

\item{n}{number of replications.}
}
\value{
{
\code{peGPD} gives the distribution function; \code{qeGPD} gives the quantile function; \code{reGPD} generates random deviates.
}
}
\description{
Distribution function, quantile function and random generation for the extended generalised Pareto distribution (eGPD) with scale \code{sigma > 0} and shapes \code{kappa > 0} and \code{xi > 0}.
Note that unlike similar functions in package \code{stats}, these functions accept only scalar inputs, rather than vectors, for the parameters.
}
\details{
{
Originally proposed by Papastathopoulos and Tawn (2013), the first family of the eGPD (as named by Naveau et al., 2016) has distribution function
\deqn{G(y|\sigma,\kappa, \xi)=[\{1+\xi(y-\mu)/\sigma\}_+^{-1/\xi}]^{\kappa}} for scale \eqn{\sigma>0} and shapes \eqn{\kappa > 0} and \eqn{\xi > 0}.

}
}
\references{
Papastathopoulos, I. and Tawn, J. A. (2013), \emph{xtended generalised Pareto models for tail estimation}, Journal of Statistical Planning and Inference, 43(1):131–1439.
(\href{https://doi.org/10.1016/j.jspi.2012.07.001}{doi})

Naveau, P., Huser, R., Ribereau, P., and Hannart, A. (2016), \emph{Modeling jointly low, moderate, and heavy rainfall intensities without a threshold selection}, Water Resources Research, 2(4):2753–2769.
(\href{https://doi.org/10.1002/2015WR018552}{doi})
}
