% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_NN.R
\name{lognormal.NN}
\alias{lognormal.NN}
\alias{lognormal.NN.train}
\alias{lognormal.NN.predict}
\title{log-normal PINN}
\usage{
lognormal.NN.train(Y.train, Y.valid = NULL, X.mu, X.sig, type = "MLP",
  link.loc = "identity", n.ep = 100, batch.size = 100, init.loc = NULL,
  init.sig = NULL, widths = c(6, 3), filter.dim = c(3, 3), seed = NULL,
  init.wb_path = NULL, S_lambda = NULL)

lognormal.NN.predict(X.mu, X.sig, model)
}
\arguments{
\item{Y.train, Y.valid}{a 2 or 3 dimensional array of training or validation real response values.
Missing values can be handled by setting corresponding entries to \code{Y.train} or \code{Y.valid} to \code{-1e10}.
The first dimension should be the observation indices, e.g., time.

If \code{type=="CNN"}, then \code{Y.train} and \code{Y.valid} must have three dimensions with the latter two corresponding to an \eqn{M} by \eqn{N} regular grid of spatial locations.
If \code{Y.valid==NULL}, no validation loss will be computed and the returned model will be that which minimises the training loss over \code{n.ep} epochs.}

\item{X.mu}{list of arrays corresponding to complementary subsets of the \eqn{d\geq 1} predictors which are used for modelling the location parameter \eqn{\mu}. Must contain at least one of the following three named entries:\describe{
\item{\code{X.lin.mu}}{A 3 or 4 dimensional array of "linear" predictor values. One more dimension than \code{Y.train}. If \code{NULL}, a model without the linear component is built and trained.
The first 2/3 dimensions should be equal to that of \code{Y.train}; the last dimension corresponds to the chosen \eqn{l_1\geq 0} 'linear' predictor values.}
\item{\code{X.add.basis.mu}}{A 4 or 5 dimensional array of basis function evaluations for the "additive" predictor values.
The first 2/3 dimensions should be equal to that of \code{Y.train}; the penultimate dimensions corresponds to the chosen \eqn{a_1\geq 0} 'linear' predictor values and the last dimension is equal to the number of knots used for estimating the splines. See example.
If \code{NULL}, a model without the additive component is built and trained.}
\item{\code{X.nn.mu}}{A 3 or 4 dimensional array of "non-additive" predictor values.  If \code{NULL}, a model without the NN component is built and trained; if this is the case, then \code{type} has no effect.
The first 2/3 dimensions should be equal to that of \code{Y.train}; the last dimension corresponds to the chosen \eqn{d-l_1-a_1\geq 0} 'non-additive' predictor values.}
}
Note that \code{X.mu} and \code{X.sig} are the predictors for both \code{Y.train} and \code{Y.valid}. If \code{is.null(X.mu)}, then \eqn{\mu} will be treated as fixed over the predictors.}

\item{X.sig}{similarly to \code{X.mu}, but for modelling the shape parameter \eqn{\sigma>0}. Note that we require at least one of \code{!is.null(X.mu)} or \code{!is.null(X.sig)}, otherwise the formulated model will be fully stationary and will not be fitted.}

\item{type}{string defining the type of network to be built. If \code{type=="MLP"}, the network will have all densely connected layers; if \code{type=="CNN"},
the network will have all convolutional layers. Defaults to an MLP. (Currently the same network is used for all parameters, may change in future versions)}

\item{n.ep}{number of epochs used for training. Defaults to 1000.}

\item{batch.size}{batch size for stochastic gradient descent. If larger than \code{dim(Y.train)[1]}, i.e., the number of observations, then regular gradient descent used.}

\item{init.loc, init.sig}{sets the initial \eqn{\mu} and \eqn{\sigma} estimates across all dimensions of \code{Y.train}. Overridden by \code{init.wb_path} if \code{!is.null(init.wb_path)}. Defaults to empirical estimates of mean and standard deviation, respectively, of \code{log(Y.train)}.}

\item{widths}{vector of widths/filters for hidden dense/convolution layers. Number of layers is equal to \code{length(widths)}. Defaults to (6,3).}

\item{filter.dim}{if \code{type=="CNN"}, this 2-vector gives the dimensions of the convolution filter kernel; must have odd integer inputs. Note that filter.dim=c(1,1) is equivalent to \code{type=="MLP"}. The same filter is applied for each hidden layer across all parameters with NN predictors.}

\item{seed}{seed for random initial weights and biases.}

\item{init.wb_path}{filepath to a \code{keras} model which is then used as initial weights and biases for training the new model. The original model must have
the exact same architecture and trained with the same input data as the new model. If \code{NULL}, then initial weights and biases are random (with seed \code{seed}) but the
final layer has zero initial weights to ensure that the initial location and shape estimates are \code{init.loc} and \code{init.sig}, respectively,  across all dimensions.}

\item{S_lambda}{List of smoothing penalty matrices for the splines modelling the effects of \code{X.add.basis.mu} and \code{X.add.basis.sig} on their respective parameters; each element only used if \code{!is.null(X.add.basis.mu)} and \code{!is.null(X.add.basis.sig)}, respectively. If \code{is.null(S_lambda[[1]])}, then no smoothing penalty used for \code{!is.null(X.add.basis.mu)}; similarly for the second element and \code{!is.null(X.add.basis.sig)}.}

\item{model}{fitted \code{keras} model. Output from \code{bGEVPP.NN.train}.}

\item{loc.link}{string defining the link function used for the location parameter, see \eqn{h_1} below. If \code{link=="exp"}, then \eqn{h_1=\exp(x)}; if \code{link=="identity"}, then \eqn{h_1(x)=x}.}
}
\value{
\code{lognormal.NN.train} returns the fitted \code{model}.  \code{lognormal.NN.predict} is a wrapper for \code{keras::predict} that returns the predicted parameter estimates, and, if applicable, their corresponding linear regression coefficients and spline bases weights.
}
\description{
Build and train a partially-interpretable neural network for fitting a log-normal model
}
\details{
{
Consider a real-valued random variable \eqn{Y} and let \eqn{\mathbf{X}} denote a \eqn{d}-dimensional predictor set with observations \eqn{\mathbf{x}}.
For \eqn{i=1,2}, we define integers \eqn{l_i\geq 0,a_i \geq 0} and \eqn{0\leq l_i+a_i \leq d}, and let \eqn{\mathbf{X}^{(i)}_L, \mathbf{X}^{(i)}_A} and \eqn{\mathbf{X}^{(i)}_N} be distinct sub-vectors
of \eqn{\mathbf{X}}, with observations of each component denoted \eqn{\mathbf{x}^{(i)}_L, \mathbf{x}^{(i)}_A} and \eqn{\mathbf{x}^{(i)}_N}, respectively; the lengths of the sub-vectors are \eqn{l_i,a_i} and \eqn{d_i-l_i-a}, respectively.
We model \eqn{Y|\mathbf{X}=\mathbf{x}\sim\mbox{Lognormal}(\mu(\mathbf{x}),\sigma(\mathbf{x}))} with \eqn{\sigma>0} and
\deqn{\mu (\mathbf{x})=h_1\{\eta^{(1)}_0+m^{(1)}_L(\mathbf{x}^{(1)}_L)+m^{(1)}_A(x^{(1)}_A)+m^{(1)}_N(\mathbf{x}^{(1)}_N)\}} and
\deqn{\sigma (\mathbf{x})=\exp\{\eta^{(2)}_0+m^{(2)}_L(\mathbf{x}^{(2)}_L)+m^{(2)}_A(x^{(2)}_A)+m^{(2)}_N(\mathbf{x}^{(2)}_N)\}}
where \eqn{h_1} is some link-function and \eqn{\eta^{(1)}_0,\eta^{(2)}_0} are constant intercepts. The unknown functions \eqn{m^{(1)}_L,m^{(2)}_L} and
\eqn{m^{(1)}_A,m^{(2)}_A} are estimated using linear functions and splines, respectively, and are
both returned as outputs by \code{lognormal.NN.predict}; \eqn{m^{(1)}_N,m^{(2)}_N} are estimated using neural networks
(currently the same architecture is used for both parameters).

For details of the log-normal parameterisation, see \code{help(Lognormal)}.

The model is fitted by minimising the negative log-likelihood associated with the lognormal distribution plus some smoothing penalty for the additive functions (determined by \code{S_lambda}; see Richards and Huser, 2022); training is performed over \code{n.ep} training epochs.
Although the model is trained by minimising the loss evaluated for \code{Y.train}, the final returned model may minimise some other loss.
The current state of the model is saved after each epoch, using \code{keras::callback_model_checkpoint}, if the value of some criterion subcedes that of the model from the previous checkpoint; this criterion is the loss evaluated for validation set \code{Y.valid} if \code{!is.null(Y.valid)} and for \code{Y.train}, otherwise.

}
}
\examples{

# Build and train a simple MLP for toy data

set.seed(1)

# Create  predictors
preds<-rnorm(prod(c(200,10,10,8)))

#Re-shape to a 4d array. First dimension corresponds to observations,
#last to the different components of the predictor set.
#Other dimensions correspond to indices of predictors, e.g., a grid of locations.
dim(preds)=c(200,10,10,8) 
#We have 200 observations of eight predictors on a 10 by 10 grid.

#Split predictors into linear, additive and nn. Different for the location and shape parameters.
X.nn.mu=preds[,,,1:4] #Four nn predictors for mu
X.lin.mu=preds[,,,5:6] #Two additive predictors for mu
X.add.mu=preds[,,,7:8] #Two additive predictors for mu

X.nn.sig=preds[,,,1:2] #Two nn predictors for sigma
X.lin.sig=preds[,,,3] #One linear predictor for sigma
dim(X.lin.sig)=c(dim(X.lin.sig),1) #Change dimension so consistent
X.add.sig=preds[,,,4] #One additive predictor for sigma
dim(X.add.sig)=c(dim(X.add.sig),1) #Change dimension so consistent


# Create toy response data

#Contribution to location parameter
#Linear contribution
m_L_1 = 0.3*X.lin.mu[,,,1]+0.6*X.lin.mu[,,,2]

# Additive contribution
m_A_1 = 0.05*X.add.mu[,,,1]^3+0.2*X.add.mu[,,,1]-
 0.05*X.add.mu[,,,2]^3+0.5*X.add.mu[,,,2]^2

#Non-additive contribution - to be estimated by NN
m_N_1 = 0.25*exp(-3+X.nn.mu[,,,4]+X.nn.mu[,,,1])+
  sin(X.nn.mu[,,,1]-X.nn.mu[,,,2])*(X.nn.mu[,,,4]+X.nn.mu[,,,2])-
  cos(X.nn.mu[,,,4]-X.nn.mu[,,,1])*(X.nn.mu[,,,3]+X.nn.mu[,,,1])

mu=m_L_1+m_A_1+m_N_1 #Identity link

#Contribution to shape parameter
#Linear contribution
m_L_2 = 0.5*X.lin.sig[,,,1]

# Additive contribution
m_A_2 = 0.1*X.add.sig[,,,1]^2+0.2*X.add.sig[,,,1]

#Non-additive contribution - to be estimated by NN
m_N_2 = 0.1*exp(-4+X.nn.sig[,,,2]+X.nn.sig[,,,1])+
  0.1* sin(X.nn.sig[,,,1]-X.nn.sig[,,,2])*(X.nn.sig[,,,1]+X.nn.sig[,,,2])

sig=0.1*exp(m_L_2+m_A_2+m_N_2) #Exponential link


theta=array(dim=c(dim(sig),2))
theta[,,,1]=mu; theta[,,,2] = sig
#We simulate data from a lognormal distribution

Y=apply(theta,1:3,function(x) rlnorm(1,meanlog =x[1],sdlog=x[2]))

#Create training and validation, respectively.
#We mask 20\% of the Y values and use this for validation
#Masked values must be set to -1e10 and are treated as missing whilst training

mask_inds=sample(1:length(Y),size=length(Y)*0.8)

Y.train<-Y.valid<-Y #Create training and validation, respectively.
Y.train[-mask_inds]=-1e10
Y.valid[mask_inds]=-1e10



#To build a model with an additive component, we require an array of evaluations of
#the basis functions for each pre-specified knot and entry to X.add.mu and X.add.sig

rad=function(x,c){ #Define a basis function. Here we use the radial bases
  out=abs(x-c)^2*log(abs(x-c))
  out[(x-c)==0]=0
 return(out)
}

n.knot.mu = 5; n.knot.sig = 4 # set number of knots.
#Must be the same for each additive predictor,
#but can differ between the parameters mu and sigma

#Get knots for mu predictors
knots.mu=matrix(nrow=dim(X.add.mu)[4],ncol=n.knot.mu)

#We set knots to be equally-spaced marginal quantiles
for( i in 1:dim(X.add.mu)[4]){
knots.mu[i,]=quantile(X.add.mu[,,,i],probs=seq(0,1,length=n.knot.mu))
}
#Evaluate radial basis functions for mu predictors
X.add.basis.mu<-array(dim=c(dim(X.add.mu),n.knot.mu))
for( i in 1:dim(X.add.mu)[4]) {
  for(k in 1:n.knot.mu) {
    X.add.basis.mu[,,,i,k]= rad(x=X.add.mu[,,,i],c=knots.mu[i,k])
    #Evaluate rad at all entries to X.add.mu and for all knots
  }}
  
  #'#Create smoothing penalty matrix for the two q_alpha additive functions

# Set smoothness parameters for two functions
 lambda = c(0.1,0.2) 

S_lambda.mu=matrix(0,nrow=n.knot.mu*dim(X.add.mu)[4],
ncol=n.knot.mu*dim(X.add.mu)[4])

for(i in 1:dim(X.add.mu)[4]){
 for(j in 1:n.knot.mu){
  for(k in 1:n.knot.mu){
     S_lambda.mu[(j+(i-1)*n.knot.mu),(k+(i-1)*n.knot.mu)]=lambda[i]*rad(knots.mu[i,j],knots.mu[i,k])
  }
}
}

#Get knots for sigma predictor

knots.sig=matrix(nrow=dim(X.add.sig)[4],ncol=n.knot.sig)

for( i in 1:dim(X.add.sig)[4]){
 knots.sig[i,]=quantile(X.add.sig[,,,i],probs=seq(0,1,length=n.knot.sig))
}

#Evaluate radial basis functions for sigma predictor

X.add.basis.sig<-array(dim=c(dim(X.add.sig),n.knot.sig))

for( i in 1:dim(X.add.sig)[4]) {
  for(k in 1:n.knot.sig) {
    X.add.basis.sig[,,,i,k]= rad(x=X.add.sig[,,,i],c=knots.sig[i,k])
    #Evaluate rad at all entries to X.add.mu and for all knots
  }}

#Create smoothing penalty matrix for the s_beta additive function

# Set smoothness parameter
 lambda = c(0.2) 

S_lambda.sig=matrix(0,nrow=n.knot.sig*dim(X.add.sig)[4],
ncol=n.knot.sig*dim(X.add.sig)[4])

for(i in 1:dim(X.add.sig)[4]){
 for(j in 1:n.knot.sig){
  for(k in 1:n.knot.sig){
     S_lambda.sig[(j+(i-1)*n.knot.sig),(k+(i-1)*n.knot.sig)]=lambda[i]*rad(knots.sig[i,j],knots.sig[i,k])
  }
}
}
 
#Join in one list
S_lambda =list("S_lambda.mu"=S_lambda.mu, "S_lambda.sig"=S_lambda.sig)


#lin+GAM+NN models defined for both location and scale parameters
X.mu=list("X.nn.mu"=X.nn.mu, "X.lin.mu"=X.lin.mu,
               "X.add.basis.mu"=X.add.basis.mu) #Predictors for mu
X.sig=list("X.nn.sig"=X.nn.sig, "X.lin.sig"=X.lin.sig,
               "X.add.basis.sig"=X.add.basis.sig) #Predictors for sigma


#Fit the log-normal model. Note that training is not run to completion.
NN.fit<-lognormal.NN.train(Y.train, Y.valid,X.mu,X.sig, type="MLP",link.loc="identity",
                          n.ep=50, batch.size=50,
                          widths=c(6,3),seed=1,S_lambda=S_lambda)
out<-lognormal.NN.predict(X.mu=X.mu,X.sig=X.sig,NN.fit$model)

print("mu linear coefficients: "); print(round(out$lin.coeff_loc,3))
print("sig linear coefficients: "); print(round(out$lin.coeff_sig,3))

# Note that this is a simple example that can be run in a personal computer. 

## To save model, run
# NN.fit$model \%>\% save_model_tf("model_lognormal")
## To load model, run
#model  <- load_model_tf("model_lognormal",
#                        custom_objects=list(
#                          "lognormal_loss_S_lambda___S_lambda_"=
#                            lognormal_loss(S_lambda=S_lambda))
#)



# Plot splines for the additive predictors

#Location predictors
n.add.preds_loc=dim(X.add.mu)[length(dim(X.add.mu))]
par(mfrow=c(1,n.add.preds_loc))
for(i in 1:n.add.preds_loc){
  plt.x=seq(from=min(knots.mu[i,]),to=max(knots.mu[i,]),length=1000)  #Create sequence for x-axis

  tmp=matrix(nrow=length(plt.x),ncol=n.knot.mu)
  for(j in 1:n.knot.mu){
    tmp[,j]=rad(plt.x,knots.mu[i,j]) #Evaluate radial basis function of plt.x and all knots
  }
  plt.y=tmp\%*\%out$gam.weights_loc[i,]
  plot(plt.x,plt.y,type="l",main=paste0("mu spline: predictor ",i),xlab="x",ylab="f(x)")
  points(knots.mu[i,],rep(mean(plt.y),n.knot.mu),col="red",pch=2)
  #Adds red triangles that denote knot locations

}

#Shape predictors
n.add.preds_sig=dim(X.add.sig)[length(dim(X.add.sig))]
par(mfrow=c(1,n.add.preds_sig))
for(i in 1:n.add.preds_sig){
  plt.x=seq(from=min(knots.sig[i,]),to=max(knots.sig[i,]),length=1000)  #Create sequence for x-axis

  tmp=matrix(nrow=length(plt.x),ncol=n.knot.sig)
  for(j in 1:n.knot.sig){
    tmp[,j]=rad(plt.x,knots.sig[i,j]) #Evaluate radial basis function of plt.x and all knots
  }
  plt.y=tmp\%*\%out$gam.weights_sig[i,]
  plot(plt.x,plt.y,type="l",main=paste0("sigma spline: predictor ",i),xlab="x",ylab="f(x)")
  points(knots.sig[i,],rep(mean(plt.y),n.knot.sig),col="red",pch=2)
  #Adds red triangles that denote knot locations

}

}
\references{
{

Richards, J. and Huser, R. (2022), \emph{A unifying partially-interpretable framework for neural network-based extreme quantile regression}. (\href{https://arxiv.org/abs/2208.07581}{arXiv:2208.07581}).
}
}
