% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedWild.R
\docType{data}
\name{MedWild}
\alias{MedWild}
\title{Mediterranean Europe Wildfire data}
\format{
A list with 6 elements:
\describe{
\item{BA}{An array with dimension (120, 200, 80), corresponding to the burnt area response data.}
\item{lambda}{An array with dimension (120, 200, 80), corresponding to the area of the spatial gridcell. Note that this is a dynamic variable.}
\item{X}{A list with 4 elements:
\describe{
\item{X.met}{An array with dimension (120, 200, 80, 13) corresponding to the meterological predictors described below.}
\item{X.lc}{An array with dimension (120, 200, 80, 21) corresponding to the land cover predictors described below.}
\item{lc.flags}{A vector of length 21 giving the names of the land cover types.}
\item{X.oro}{An array with dimension (120, 200, 80, 2) corresponding to the oropgraphical predictors described below.}

}
}
\item{times}{A vector of length 120 giving the observation indices. Format is "year-month". Corresponds to first dimension of \code{BA}.}
\item{lon}{A matrix of dimension (200, 80) giving the longitude ordinate for the second and third dimensions of \code{BA}.}
\item{lat}{A matrix of dimension (200, 80) giving the latitude ordinate for the second and third dimensions of \code{BA}.}
}
}
\usage{
data(MedWild)
}
\description{
Data used by Richards et al. (2022) for modelling extreme wildfires in Mediterranean Europe
with partially-interpretable neural networks
}
\details{
The response data \code{BA} are observations of monthly aggregated burnt area (km^2) for over 10000 spatial grid-cells
located across countries surrounding the Mediterranean Basin and southern Europe.
The observation period covers 2001 to 2020, using only months between June and November, inclusive,
leaving 120 observed spatial fields. Grid-cells are arranged on a regular latitude/longitude grid with spatial resolution
0.25deg by 0.25deg. Observations are derived from version 5.1 of the FireCCI dataset (Lizundia-Loiola et al., 2020),
which is generated by the Moderate Resolution Imaging Spectroradiometer (MODIS) 250m reflectance data
and guided by 1km active fires (Otón et al., 2021). Alongside values of BA, we provide a measure of burnable-area for a space-time grid-cell, given in \code{lambda}; this is equivalent to \eqn{\lambda(s,t)} in Richards et al., 2022, and can be specified as an offset term in both \link{GPD.NN} and \link{quant.NN}.

Both the response data and the subsequently described predictors have been re-gridded to a regular spatio-temporal grid with missing values set to \code{-1e10}.
For \code{BA} and entries to \code{X}, the first three dimensions correspond to time \eqn{\times} latitude \eqn{\times} longitude with
their respective ordinate values given in \code{times}, \code{lat} and \code{lon}.

We consider three types of predictor variables, given in \code{X}: these are meteorological (\code{X.met}), land cover proportions (\code{X.lc}) and
orograhical (\code{X.oro}).

Thirteen meteorological variables are considered and given as monthly means in \code{X.met}.
Eleven variables are provided by the ERA5-reanalysis on single levels (Hersbach et al., 2019), available through the COPERNICUS Climate Data Service. Monthly total precipitation (m) is used to derive a three-month standardized precipitation index (SPI, unitless) and hourly temperature and dewpoint temperature at a 2m altitude (K) are used to derive monthly vapour-pressure deficit (VPD, measured in Pa).
The variables are ordered as followed: VPD (Pa), temperature at a 2m altitude (K), SPI (unitless), both eastern and northern components of wind velocity at a 10m altitude (m/s),
potential evaporation (m), surface pressure (Pa), evaporation (m of water equiv.), surface net thermal, and solar, radiation (J/m^2),
snow evaporation (m of water equiv.), snowmelt (m of water equiv.) and snowfall (m of water equiv.)

The land cover variables that are given in \code{X.lc} describe the proportion of a grid-cell which
is covered by one of 21 different types, e.g., urban, grassland, water.
Land cover predictors are derived using a gridded land cover map, of spatial resolution 300m and temporal resolution one year,
produced by COPERNICUS and available through their Climate Data Service. For each 0.25deg \eqn{\times} 0.25deg spatial grid-cell,
the proportion of a cell consisting of a specific land cover type is derived from the high-resolution product.
The variables are named in \code{lc.flags}.

The two orographical predictors given in \code{X.oro} are the mean and standard deviation of the altitude (m)
for each grid-cell; estimates are derived using the R package ``elevatr",
which accesses Amazon Web Services Terrain Tiles (\url{https://registry.opendata.aws/terrain-tiles/}).
}
\examples{
data(MedWild)
}
\references{
Bontemps, S., Boettcher, M., Brockmann, C., Kirches, G., Lamarche, C., Radoux, J., Santoro, M., Vanbogaert, E., Wegmüller, U., Herold, M., Achard, F., Ramoino, F., Arino,O., and Defourny, P. (2015). \emph{Multi-year global land cover mapping at 300 M and characterization for climate modelling: Achievements of the land cover component of the ESA climate change initiative.} The International Archives of Photogrammetry, Remote Sensing and Spatial Information Sciences, 40(7):323.(\href{https://doi.org/10.5194/isprsarchives-XL-7-W3-323-2015}{doi})

Hersbach, H., Bell, B., Berrisford, P., Biavati, G., Horányi, A., Muñoz Sabater, J., Nicola s,J., Peubey, C., Radu, R., Rozum, I., Schepers, D., Simmons, A., Soci, C., Dee, D., and Tépaut, J.-N. (2019). \emph{ERA5 monthly averaged data on single levels from 1959 to present.} Copernicus Climate Change Service (C3S) Climate Data Store (CDS), 10:252–266. (\href{https://doi.org/10.24381/cds.f17050d7}{doi})

Lizundia-Loiola, J., Otón, G., Ramo, R., and Chuvieco, E. (2020). \emph{A spatio-temporal active-fire clustering approach for global burned area mapping at 250 m from MODIS data.} Remote Sensing of Environment, 236:111493. (\href{https://doi.org/10.1016/j.rse.2019.111493}{doi})

Otón, G., Pereira, J. M. C., Silva, J. M. N., and Chuvieco, E. (2021). \emph{Analysis of trends in the FireCCI global long term burned area product (1982–2018).} Fire, 4(4). (\href{https://doi.org/10.3390/fire4040074}{doi})

Richards, J., Huser, R., Bevacqua, E., Zscheischler, J,  (2022), \emph{Insights into the drivers and spatio-temporal trends of extreme Mediterranean wildfires with statistical deep-learning.} (\href{https://arxiv.org/abs/2212.01796}{arXiv:2212.01796})
}
\keyword{datasets}
